--- a/spacy/syntax/nn_parser.pyx
+++ b/spacy/syntax/nn_parser.pyx
@@ -271,7 +271,8 @@ cdef class Parser:
 
         # TODO: This is an unfortunate hack atm!
         # Used to set input dimensions in network.
-        lower.begin_training(lower.ops.allocate((500, token_vector_width)))
+        if not cfg.get('from_disk', False):
+            lower.begin_training(lower.ops.allocate((500, token_vector_width)))
         cfg = {
             'nr_class': nr_class,
             'hidden_depth': depth,
@@ -888,7 +889,7 @@ cdef class Parser:
             path = util.ensure_path(path)
             if self.model is True:
                 self.cfg['pretrained_dims'] = self.vocab.vectors_length
-                self.model, cfg = self.Model(**self.cfg)
+                self.model, cfg = self.Model(from_disk=True, **self.cfg)
             else:
                 cfg = {}
             with (path / 'tok2vec_model').open('rb') as file_: