Performance of loading a model from disk was too slow. After profiling, I found that the issue was random-initialization of weights which will be overwritten by data from the disk. This patch improves the performance of load from disk by more than 5x.

## Description
After doing performance profiling on loading models from disk, I found out that each model is initialized with some random weights and then those weights are overwritten by data from disk. This overhead of random initialization was ~80% of total load time.

In my test that loads 38 different models. Below is the profiling of that script without this patch:
`         2897440 function calls (2730892 primitive calls) in 94.368 seconds

   Ordered by: internal time
   List reduced from 3254 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      228   37.318    0.164   37.318    0.164 {built-in method numpy.core.multiarray.dot}
      228   28.010    0.123   74.568    0.327 _ml.py:198(predict)
      228    8.025    0.035    8.025    0.035 {built-in method numpy.core.multiarray.concatenate}
       38    3.777    0.099    4.491    0.118 {method 'from_disk' of 'spacy.vocab.Vocab' objects}
     1786    3.429    0.002    3.429    0.002 {method 'uniform' of 'mtrand.RandomState' objects}
       76    2.439    0.032   79.594    1.047 _ml.py:175(init_weights)
       76    2.026    0.027    2.026    0.027 {method 'normal' of 'mtrand.RandomState' objects}
       38    1.291    0.034    1.647    0.043 language.py:54(create_tokenizer)
      228    1.113    0.005   46.523    0.204 _ml.py:131(begin_update)
       38    1.023    0.027    1.148    0.030 {method 'from_disk' of 'spacy.tokenizer.Tokenizer' objects}
`
The script's total time was 94.36 seconds.

After small modification of how the Model is initialized, the profiling of the script is below:
`         2881472 function calls (2715159 primitive calls) in 15.252 seconds

   Ordered by: internal time
   List reduced from 3227 to 10 due to restriction <10>


   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       38    3.832    0.101    4.580    0.121 {method 'from_disk' of 'spacy.vocab.Vocab' objects}
     1710    3.368    0.002    3.368    0.002 {method 'uniform' of 'mtrand.RandomState' objects}
       38    1.254    0.033    1.604    0.042 language.py:54(create_tokenizer)
       76    1.107    0.015    4.950    0.065 nn_parser.pyx:856(from_disk)
       38    0.996    0.026    1.092    0.029 {method 'from_disk' of 'spacy.tokenizer.Tokenizer' objects}
       38    0.667    0.018    0.667    0.018 _ml.py:228(link_vectors_to_models)
     3800    0.499    0.000    0.501    0.000 util.py:46(copy_array)
       38    0.432    0.011    2.271    0.060 pipeline.pyx:580(load_model)
      418    0.305    0.001    0.556    0.001 {msgpack._unpacker.unpackb}
      456    0.229    0.001    3.362    0.007 hash_embed.py:15(wrapped)
`
Total runtime of the test reduced to 15.25 seconds - which is more than 6x performance improvement.

### Types of change
It's a performance enhancement in loading a model from disk.

## Checklist

- [X] I have submitted the spaCy Contributor Agreement.
- [X] I ran the tests, and all new and existing tests passed.
- [X] My changes don't require a change to the documentation, or if they do, I've added all required information.