@honnibal 2018-01-17
Thanks for pointing this out!

I think we should be able to make this a little cleaner. What if we moved the call to `begin_training()` out of the `Model` method entirely? Then we don't have to introduce this `from_disk` parameter -- which also doesn't cover loading from byte strings with `from_bytes`.

From limited tests, the following patch appears to work:

```
--- a/spacy/syntax/nn_parser.pyx
+++ b/spacy/syntax/nn_parser.pyx
@@ -269,9 +269,6 @@ cdef class Parser:
                 zero_init(Affine(nr_class, hidden_width, drop_factor=0.0))
             )
 
-        # TODO: This is an unfortunate hack atm!
-        # Used to set input dimensions in network.
-        lower.begin_training(lower.ops.allocate((500, token_vector_width)))
         cfg = {
             'nr_class': nr_class,
             'hidden_depth': depth,
@@ -840,6 +837,8 @@ cdef class Parser:
             self.cfg.update(cfg)
         elif sgd is None:
             sgd = self.create_optimizer()
+        self.model[1].begin_training(
+            self.model[1].ops.allocate((5, cfg['token_vector_width'])))
         return sgd
 
     def init_multitask_objectives(self, gold_tuples, pipeline, **cfg):
```