@avadhpatel 2018-01-18
I think that should work. Can you update the code or want me to resubmit the patch? I am ok with either.