@honnibal 2018-01-20
@avadhpatel If you get to it this weekend, it'd be nice to have the pull request be "yours" so you can be credited correctly. If you'd rather not bother, I'm happy to make the patch.

Thanks again!