@avadhpatel 2018-01-21
Ok I'll send a new pull request with suggested changes soon.